% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{show,BQCResult-method}
\alias{show,BQCResult-method}
\title{Show method for BQCResult}
\usage{
\S4method{show}{BQCResult}(object)
}
\arguments{
\item{object}{A \code{BQCResult} object.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Prints a compact summary of a \code{BQCResult} object.
}
