% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BQCResult}
\alias{BQCResult}
\alias{show,BQCResult-method}
\alias{qcFlags,BQCResult-method}
\alias{doubletScores,BQCResult-method}
\alias{batchSummary,BQCResult-method}
\title{Constructor for BQCResult}
\usage{
BQCResult(qcFlags, doubletScores, batchSummary, params = list())

\S4method{show}{BQCResult}(object)

\S4method{qcFlags}{BQCResult}(x, ...)

\S4method{doubletScores}{BQCResult}(x, ...)

\S4method{batchSummary}{BQCResult}(x, ...)
}
\arguments{
\item{qcFlags}{A \code{DataFrame} of per-cell QC flags.}

\item{doubletScores}{A numeric vector of doublet scores.}

\item{batchSummary}{A \code{DataFrame} of per-batch statistics.}

\item{params}{A list of analysis parameters.}
}
\value{
A \code{BQCResult} object.
}
\description{
Create a new \code{BQCResult} object.
}
